Use stdint types and add MOTOR_CHANNELS
Replace native Arduino types with fixed-width stdint types (uint8_t, uint16_t, int16_t, uint32_t) and add #include <stdint.h> across headers to improve type safety and portability. Introduce MOTOR_CHANNELS defines and a default MOTOR_CHANNELS = 0 fallback in Config.h; set MOTOR_CHANNELS for relevant configurations and use it to index servo outputs. Tighten function signatures and variables (const pin parameter, change packet/timing types to uint32_t, constrain math casts), replace byte/int loop counters with uint8_t/uint16_t where appropriate, and consolidate servo indexing to account for motor channels. Changes span Config.h, Motor_Control.ino, PWM.h, Pins.h, RSSI.h/.ino, RX.h/.ino, RX_TX_Util.h/.ino, and Servo_Control.ino to reduce warnings and improve correctness.

+++ RX_nRF24_Cabell/RSSI.h
@@ -28,6 +28,7 @@
 #define __RSSI_h__
 
 #include <Arduino.h>
+#include <stdint.h>
 
 #define RSSI_CALC_INTERVAL             152 // The number of expected packets between calculating the base RSSI. Approximately 1/2 second
 #define TELEMETRY_RSSI_MIN_VALUE       0   // The lowest possible RSSI value = zero packet rate
@@ -52,13 +53,13 @@ class RSSI
   void packetProcess();
   void resetCounters();
   
-  int hitCount = 0;
-  int missCount = 0;
-  int secondaryHitCount = 0;
-  int badPacketCount = 0;
-  int packetCount = 0;
+  int16_t hitCount = 0;
+  int16_t missCount = 0;
+  int16_t secondaryHitCount = 0;
+  int16_t badPacketCount = 0;
+  int16_t packetCount = 0;
 
-  int packetRate = 0.0;
+  int16_t packetRate = 0.0;
   int8_t sequentialMissTrack = 0;
 };
